refactor : pool 인터페이스를 types 폴더에 있는 걸로 적용
