+++ src/types/coordinate.ts
@@ -1,0 +1,4 @@
+export interface Coordinates {
+  lat: number
+  lng: number
+}
